EIP1-5628, EIP1-5662 - Update mappers and remove mention of AED "summary"
