+++ src/main/kotlin/uk/gov/dluhc/printapi/service/aed/AnonymousElectorDocumentService.kt
@@ -6,11 +6,11 @@ import uk.gov.dluhc.printapi.client.ElectoralRegistrationOfficeNotFoundException
 import uk.gov.dluhc.printapi.database.entity.SourceType.ANONYMOUS_ELECTOR_DOCUMENT
 import uk.gov.dluhc.printapi.database.repository.AnonymousElectorDocumentRepository
 import uk.gov.dluhc.printapi.dto.PdfFile
-import uk.gov.dluhc.printapi.dto.aed.AnonymousElectorDocumentSummaryDto
+import uk.gov.dluhc.printapi.dto.aed.AnonymousElectorDocumentDto
 import uk.gov.dluhc.printapi.dto.aed.GenerateAnonymousElectorDocumentDto
 import uk.gov.dluhc.printapi.exception.GenerateAnonymousElectorDocumentValidationException
 import uk.gov.dluhc.printapi.mapper.aed.AnonymousElectorDocumentMapper
-import uk.gov.dluhc.printapi.mapper.aed.AnonymousElectorSummaryMapper
+import uk.gov.dluhc.printapi.mapper.aed.GenerateAnonymousElectorDocumentMapper
 import uk.gov.dluhc.printapi.service.EroService
 import uk.gov.dluhc.printapi.service.pdf.PdfFactory
 
@@ -18,8 +18,8 @@ import uk.gov.dluhc.printapi.service.pdf.PdfFactory
 class AnonymousElectorDocumentService(
     private val eroService: EroService,
     private val anonymousElectorDocumentRepository: AnonymousElectorDocumentRepository,
+    private val generateAnonymousElectorDocumentMapper: GenerateAnonymousElectorDocumentMapper,
     private val anonymousElectorDocumentMapper: AnonymousElectorDocumentMapper,
-    private val anonymousElectorSummaryMapper: AnonymousElectorSummaryMapper,
     private val pdfTemplateDetailsFactory: AedPdfTemplateDetailsFactory,
     private val pdfFactory: PdfFactory
 ) {
@@ -28,7 +28,7 @@ class AnonymousElectorDocumentService(
     fun generateAnonymousElectorDocument(eroId: String, dto: GenerateAnonymousElectorDocumentDto): PdfFile {
         verifyGssCodeIsValidForEro(eroId, dto.gssCode)
         val filename = pdfTemplateDetailsFactory.getTemplateFilename(dto.gssCode)
-        with(anonymousElectorDocumentMapper.toAnonymousElectorDocument(dto, filename)) {
+        with(generateAnonymousElectorDocumentMapper.toAnonymousElectorDocument(dto, filename)) {
             val templateDetails = pdfTemplateDetailsFactory.getTemplateDetails(this)
             val contents = pdfFactory.createPdfContents(templateDetails)
             anonymousElectorDocumentRepository.save(this)
@@ -36,17 +36,17 @@ class AnonymousElectorDocumentService(
         }
     }
 
-    fun getAnonymousElectorDocumentSummaries(
+    fun getAnonymousElectorDocuments(
         eroId: String,
         applicationId: String
-    ): List<AnonymousElectorDocumentSummaryDto> {
+    ): List<AnonymousElectorDocumentDto> {
         val gssCodes = eroService.lookupGssCodesForEro(eroId)
         return anonymousElectorDocumentRepository.findByGssCodeInAndSourceTypeAndSourceReference(
             gssCodes = gssCodes,
             sourceType = ANONYMOUS_ELECTOR_DOCUMENT,
             sourceReference = applicationId
         ).sortedByDescending { it.dateCreated }
-            .map { anonymousElectorSummaryMapper.mapToAnonymousElectorDocumentSummaryDto(it) }
+            .map { anonymousElectorDocumentMapper.mapToAnonymousElectorDocumentDto(it) }
     }
 
     private fun verifyGssCodeIsValidForEro(eroId: String, gssCode: String) {
