+++ src/main/resources/openapi/PrintAPIs.yaml
@@ -1,7 +1,7 @@
 openapi: 3.0.0
 info:
   title: Print APIs
-  version: '1.14.0'
+  version: '1.15.0'
   description: Print APIs
   contact:
     name: Krister Bone
