+++ src/test/kotlin/uk/gov/dluhc/printapi/mapper/aed/AnonymousElectorDocumentMapperTest.kt
@@ -4,66 +4,46 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.NullAndEmptySource
+import org.junit.jupiter.params.provider.ValueSource
 import org.mockito.InjectMocks
 import org.mockito.Mock
-import org.mockito.Spy
 import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.kotlin.any
 import org.mockito.kotlin.given
 import org.mockito.kotlin.verify
+import org.mockito.kotlin.verifyNoInteractions
 import org.mockito.kotlin.verifyNoMoreInteractions
-import uk.gov.dluhc.printapi.database.entity.AedContactDetails
-import uk.gov.dluhc.printapi.database.entity.AnonymousElectorDocument
-import uk.gov.dluhc.printapi.database.entity.AnonymousElectorDocumentStatus
-import uk.gov.dluhc.printapi.database.entity.CertificateLanguage
-import uk.gov.dluhc.printapi.database.entity.DeliveryAddressType
-import uk.gov.dluhc.printapi.database.entity.SourceType.ANONYMOUS_ELECTOR_DOCUMENT
-import uk.gov.dluhc.printapi.dto.AddressFormat
-import uk.gov.dluhc.printapi.dto.DeliveryAddressType.ERO_COLLECTION
-import uk.gov.dluhc.printapi.dto.DeliveryClass
-import uk.gov.dluhc.printapi.dto.SourceType
-import uk.gov.dluhc.printapi.dto.aed.GenerateAnonymousElectorDocumentDto
+import uk.gov.dluhc.printapi.database.entity.DeliveryAddressType.REGISTERED
+import uk.gov.dluhc.printapi.database.entity.SupportingInformationFormat
 import uk.gov.dluhc.printapi.mapper.CertificateLanguageMapper
 import uk.gov.dluhc.printapi.mapper.DeliveryAddressTypeMapper
-import uk.gov.dluhc.printapi.mapper.SourceTypeMapper
-import uk.gov.dluhc.printapi.models.DeliveryAddressType.ERO_MINUS_COLLECTION
-import uk.gov.dluhc.printapi.models.SourceType.ANONYMOUS_MINUS_ELECTOR_MINUS_DOCUMENT
-import uk.gov.dluhc.printapi.service.IdFactory
-import uk.gov.dluhc.printapi.testsupport.testdata.aValidAnonymousElectorDocumentTemplateFilename
-import uk.gov.dluhc.printapi.testsupport.testdata.aValidUserId
-import uk.gov.dluhc.printapi.testsupport.testdata.aValidVacNumber
-import uk.gov.dluhc.printapi.testsupport.testdata.dto.aed.buildGenerateAnonymousElectorDocumentDto
+import uk.gov.dluhc.printapi.mapper.InstantMapper
+import uk.gov.dluhc.printapi.models.AnonymousElectorDocumentStatus
+import uk.gov.dluhc.printapi.models.CertificateLanguage
+import uk.gov.dluhc.printapi.models.DeliveryAddressType
+import uk.gov.dluhc.printapi.testsupport.testdata.aValidGeneratedDateTime
+import uk.gov.dluhc.printapi.testsupport.testdata.dto.aed.buildAnonymousElectorDocumentDto
+import uk.gov.dluhc.printapi.testsupport.testdata.dto.aed.buildAnonymousElectorDto
 import uk.gov.dluhc.printapi.testsupport.testdata.dto.aed.buildValidAddressDto
-import uk.gov.dluhc.printapi.testsupport.testdata.dto.buildDtoCertificateDelivery
 import uk.gov.dluhc.printapi.testsupport.testdata.entity.buildAddress
-import uk.gov.dluhc.printapi.testsupport.testdata.entity.buildDelivery
-import uk.gov.dluhc.printapi.testsupport.testdata.model.buildApiCertificateDelivery
-import uk.gov.dluhc.printapi.testsupport.testdata.model.buildGenerateAnonymousElectorDocumentRequest
-import java.time.Clock
-import java.time.Instant
-import java.time.LocalDate
-import java.time.ZoneOffset
-import uk.gov.dluhc.printapi.database.entity.CertificateLanguage as CertificateLanguageEntity
-import uk.gov.dluhc.printapi.database.entity.SupportingInformationFormat as SupportingInformationFormatEntity
-import uk.gov.dluhc.printapi.dto.aed.AnonymousSupportingInformationFormat as AnonymousSupportingInformationFormatDto
-import uk.gov.dluhc.printapi.models.AnonymousSupportingInformationFormat as AnonymousSupportingInformationFormatApi
+import uk.gov.dluhc.printapi.testsupport.testdata.entity.buildAedContactDetails
+import uk.gov.dluhc.printapi.testsupport.testdata.entity.buildAnonymousElectorDocument
+import uk.gov.dluhc.printapi.testsupport.testdata.model.buildAnonymousElector
+import uk.gov.dluhc.printapi.testsupport.testdata.model.buildAnonymousElectorDocument
+import uk.gov.dluhc.printapi.testsupport.testdata.model.buildValidAddress
+import uk.gov.dluhc.printapi.dto.CertificateLanguage as DtoCertificateLanguage
+import uk.gov.dluhc.printapi.dto.DeliveryAddressType as DtoDeliveryAddressType
+import uk.gov.dluhc.printapi.dto.aed.AnonymousElectorDocumentStatus as DtoAnonymousElectorDocumentStatus
+import uk.gov.dluhc.printapi.dto.aed.AnonymousSupportingInformationFormat as AnonymousSupportingInformationFormatDtoEnum
+import uk.gov.dluhc.printapi.models.AnonymousSupportingInformationFormat as AnonymousSupportingInformationFormatApiEnum
 
 @ExtendWith(MockitoExtension::class)
 class AnonymousElectorDocumentMapperTest {
 
-    companion object {
-        private const val FIXED_DATE_STRING = "2022-10-18"
-        private val FIXED_DATE = LocalDate.parse(FIXED_DATE_STRING)
-        private val FIXED_TIME = Instant.parse("${FIXED_DATE_STRING}T11:22:32.123Z")
-        private val FIXED_CLOCK = Clock.fixed(FIXED_TIME, ZoneOffset.UTC)
-        private val ID_FIELDS_REGEX = ".*id"
-    }
-
     @InjectMocks
     private lateinit var mapper: AnonymousElectorDocumentMapperImpl
-
-    @Mock
-    private lateinit var sourceTypeMapper: SourceTypeMapper
 
     @Mock
     private lateinit var certificateLanguageMapper: CertificateLanguageMapper
@@ -75,73 +55,37 @@ class AnonymousElectorDocumentMapperTest {
     private lateinit var deliveryAddressTypeMapper: DeliveryAddressTypeMapper
 
     @Mock
-    private lateinit var idFactory: IdFactory
-
-    @Spy
-    private val clock: Clock = FIXED_CLOCK
+    private lateinit var instantMapper: InstantMapper
 
     @Nested
-    inner class ToAnonymousElectorDocument {
-
+    inner class MapToApiAnonymousElectorDocument {
         @Test
-        fun `should map GenerateAnonymousElectorDocumentDto to an AnonymousElectorDocument entity`() {
+        fun `should map AnonymousElectorDocumentDto to an AnonymousElectorDocument API model`() {
             // Given
-            val dtoRequest =
-                buildGenerateAnonymousElectorDocumentDto(delivery = buildDtoCertificateDelivery(deliveryAddressType = ERO_COLLECTION))
-            val aedTemplateFilename = aValidAnonymousElectorDocumentTemplateFilename()
-            val certificateNumber = aValidVacNumber()
+            val dtoRequest = buildAnonymousElectorDocumentDto(supportingInformationFormat = AnonymousSupportingInformationFormatDtoEnum.EASY_READ)
+            val requestDateTime = aValidGeneratedDateTime()
 
-            given(sourceTypeMapper.mapDtoToEntity(any())).willReturn(ANONYMOUS_ELECTOR_DOCUMENT)
-            given(certificateLanguageMapper.mapDtoToEntity(any())).willReturn(CertificateLanguage.EN)
-            given(supportingInformationFormatMapper.mapDtoToEntity(any())).willReturn(SupportingInformationFormatEntity.EASY_READ)
-            given(deliveryAddressTypeMapper.mapDtoToEntity(any())).willReturn(DeliveryAddressType.ERO_COLLECTION)
-            given(idFactory.vacNumber()).willReturn(certificateNumber)
+            given(certificateLanguageMapper.mapDtoToApi(any())).willReturn(CertificateLanguage.EN)
+            given(supportingInformationFormatMapper.mapDtoToApi(any())).willReturn(AnonymousSupportingInformationFormatApiEnum.EASY_MINUS_READ)
+            given(deliveryAddressTypeMapper.mapDtoToApi(any())).willReturn(DeliveryAddressType.REGISTERED)
+            given(instantMapper.toOffsetDateTime(any())).willReturn(requestDateTime)
 
             val expected = with(dtoRequest) {
-                AnonymousElectorDocument(
-                    id = null,
-                    gssCode = gssCode,
-                    sourceType = ANONYMOUS_ELECTOR_DOCUMENT,
-                    sourceReference = sourceReference,
-                    applicationReference = applicationReference,
-                    certificateLanguage = CertificateLanguageEntity.EN,
-                    supportingInformationFormat = SupportingInformationFormatEntity.EASY_READ,
+                buildAnonymousElectorDocument(
                     certificateNumber = certificateNumber,
-                    photoLocationArn = photoLocation,
                     electoralRollNumber = electoralRollNumber,
-                    issueDate = FIXED_DATE,
-                    aedTemplateFilename = aedTemplateFilename,
-                    requestDateTime = FIXED_TIME,
-                    contactDetails = AedContactDetails(
-                        firstName = firstName,
-                        middleNames = middleNames,
-                        surname = surname,
-                        email = email,
-                        phoneNumber = phoneNumber,
-                        address = with(registeredAddress) {
-                            buildAddress(
-                                street = street,
-                                postcode = postcode,
-                                property = property,
-                                locality = locality,
-                                town = town,
-                                area = area,
-                                uprn = uprn
-                            )
-                        },
-                    ),
-                    userId = userId,
-                    statusHistory = mutableListOf(
-                        AnonymousElectorDocumentStatus(
-                            status = AnonymousElectorDocumentStatus.Status.PRINTED,
-                            eventDateTime = FIXED_TIME
-                        )
-                    ),
-                    delivery = with(delivery) {
-                        buildDelivery(
+                    gssCode = gssCode,
+                    certificateLanguage = CertificateLanguage.EN,
+                    supportingInformationFormat = AnonymousSupportingInformationFormatApiEnum.EASY_MINUS_READ,
+                    deliveryAddressType = DeliveryAddressType.REGISTERED,
+                    elector = with(elector) {
+                        buildAnonymousElector(
+                            firstName = firstName,
+                            middleNames = middleNames,
+                            surname = surname,
                             addressee = addressee,
-                            address = with(deliveryAddress) {
-                                buildAddress(
+                            registeredAddress = with(registeredAddress) {
+                                buildValidAddress(
                                     street = street,
                                     postcode = postcode,
                                     property = property,
@@ -151,177 +95,176 @@ class AnonymousElectorDocumentMapperTest {
                                     uprn = uprn,
                                 )
                             },
-                            deliveryClass = uk.gov.dluhc.printapi.database.entity.DeliveryClass.STANDARD,
-                            deliveryAddressType = DeliveryAddressType.ERO_COLLECTION,
-                            addressFormat = uk.gov.dluhc.printapi.database.entity.AddressFormat.UK,
+                            email = email,
+                            phoneNumber = phoneNumber,
                         )
-                    }
-
+                    },
+                    photoLocation = photoLocationArn,
+                    issueDate = issueDate,
+                    status = AnonymousElectorDocumentStatus.PRINTED,
+                    userId = userId,
+                    dateTime = requestDateTime,
                 )
             }
 
             // When
-            val actual = mapper.toAnonymousElectorDocument(dtoRequest, aedTemplateFilename)
+            val actual = mapper.mapToApiAnonymousElectorDocument(dtoRequest)
 
             // Then
-            assertThat(actual).usingRecursiveComparison().ignoringFieldsMatchingRegexes(ID_FIELDS_REGEX).isEqualTo(expected)
-            verify(sourceTypeMapper).mapDtoToEntity(dtoRequest.sourceType)
-            verify(certificateLanguageMapper).mapDtoToEntity(dtoRequest.certificateLanguage)
-            verify(supportingInformationFormatMapper).mapDtoToEntity(dtoRequest.supportingInformationFormat)
-            verify(idFactory).vacNumber()
-            verify(deliveryAddressTypeMapper).mapDtoToEntity(ERO_COLLECTION)
-            verifyNoMoreInteractions(sourceTypeMapper, certificateLanguageMapper, supportingInformationFormatMapper, idFactory, deliveryAddressTypeMapper)
+            assertThat(actual).usingRecursiveComparison().isEqualTo(expected)
+            verify(certificateLanguageMapper).mapDtoToApi(dtoRequest.certificateLanguage)
+            verify(supportingInformationFormatMapper).mapDtoToApi(dtoRequest.supportingInformationFormat)
+            verify(deliveryAddressTypeMapper).mapDtoToApi(DtoDeliveryAddressType.REGISTERED)
+            verify(instantMapper).toOffsetDateTime(dtoRequest.requestDateTime)
+            verifyNoMoreInteractions(
+                certificateLanguageMapper, supportingInformationFormatMapper,
+                deliveryAddressTypeMapper, instantMapper
+            )
         }
     }
 
     @Nested
-    inner class ToAedContactDetailsEntity {
+    inner class MapToAnonymousElectorDocumentDto {
+        @Test
+        fun `should map AnonymousElectorDocument entity to an AnonymousElectorDocumentDto`() {
+            // Given
+            val entityRequest = buildAnonymousElectorDocument(
+                supportingInformationFormat = SupportingInformationFormat.BRAILLE,
+                contactDetails = buildAedContactDetails(
+                    firstName = "John",
+                    middleNames = "J",
+                    surname = "Bloggs"
+                )
+            )
+            given(certificateLanguageMapper.mapEntityToDto(any())).willReturn(DtoCertificateLanguage.EN)
+            given(supportingInformationFormatMapper.mapEntityToDto(any())).willReturn(AnonymousSupportingInformationFormatDtoEnum.BRAILLE)
+            given(deliveryAddressTypeMapper.mapEntityToDto(any())).willReturn(DtoDeliveryAddressType.REGISTERED)
+
+            val expected = with(entityRequest) {
+                buildAnonymousElectorDocumentDto(
+                    certificateNumber = certificateNumber,
+                    electoralRollNumber = electoralRollNumber,
+                    gssCode = gssCode,
+                    certificateLanguage = DtoCertificateLanguage.EN,
+                    supportingInformationFormat = AnonymousSupportingInformationFormatDtoEnum.BRAILLE,
+                    deliveryAddressType = DtoDeliveryAddressType.REGISTERED,
+                    elector = with(contactDetails!!) {
+                        buildAnonymousElectorDto(
+                            firstName = firstName,
+                            middleNames = middleNames,
+                            surname = surname,
+                            addressee = "John J Bloggs",
+                            registeredAddress = with(address!!) {
+                                buildValidAddressDto(
+                                    street = street!!,
+                                    postcode = postcode!!,
+                                    property = property,
+                                    locality = locality,
+                                    town = town,
+                                    area = area,
+                                    uprn = uprn,
+                                )
+                            },
+                            email = email,
+                            phoneNumber = phoneNumber
+                        )
+                    },
+                    photoLocationArn = photoLocationArn,
+                    issueDate = issueDate,
+                    status = DtoAnonymousElectorDocumentStatus.PRINTED,
+                    userId = userId,
+                    requestDateTime = requestDateTime,
+                )
+            }
+
+            // When
+            val actual = mapper.mapToAnonymousElectorDocumentDto(entityRequest)
+
+            // Then
+            assertThat(actual).usingRecursiveComparison().isEqualTo(expected)
+            verify(certificateLanguageMapper).mapEntityToDto(entityRequest.certificateLanguage)
+            verify(supportingInformationFormatMapper).mapEntityToDto(entityRequest.supportingInformationFormat!!)
+            verify(deliveryAddressTypeMapper).mapEntityToDto(REGISTERED)
+            verifyNoMoreInteractions(certificateLanguageMapper, supportingInformationFormatMapper, deliveryAddressTypeMapper)
+            verifyNoInteractions(instantMapper)
+        }
+    }
+
+    @Nested
+    inner class MapFromContactDetailsToElectorDto {
 
         @Test
-        fun `should map to AedContactDetails entity given GenerateAnonymousElectorDocumentDto`() {
+        fun `should map AedContactDetails entity to AnonymousElectorDto given all names present`() {
             // Given
-            val request = buildGenerateAnonymousElectorDocumentDto()
-            val expected = with(request) {
-                AedContactDetails(
+            val aedContactDetailsEntity =
+                buildAedContactDetails(firstName = "Joe", middleNames = "John", surname = "Bloggs")
+            val expected = with(aedContactDetailsEntity) {
+                buildAnonymousElectorDto(
                     firstName = firstName,
                     middleNames = middleNames,
                     surname = surname,
-                    email = email,
-                    phoneNumber = phoneNumber,
-                    address = with(registeredAddress) {
-                        buildAddress(
-                            street = street,
-                            postcode = postcode,
-                            property = property,
-                            locality = locality,
-                            town = town,
-                            area = area,
-                            uprn = uprn
-                        )
-                    },
-                )
-            }
-
-            // When
-            val actual = mapper.toAedContactDetailsEntity(request)
-
-            // Then
-            assertThat(actual).usingRecursiveComparison().ignoringFieldsMatchingRegexes(ID_FIELDS_REGEX).isEqualTo(expected)
-        }
-    }
-
-    @Nested
-    inner class FromDeliveryDtoToDeliveryEntity {
-
-        @Test
-        fun `should map to Delivery entity given CertificateDelivery Dto`() {
-            // Given
-            val dto = buildDtoCertificateDelivery(deliveryAddressType = ERO_COLLECTION)
-            given(deliveryAddressTypeMapper.mapDtoToEntity(any())).willReturn(DeliveryAddressType.ERO_COLLECTION)
-
-            val expected = with(dto) {
-                buildDelivery(
-                    addressee = addressee,
-                    addressFormat = uk.gov.dluhc.printapi.database.entity.AddressFormat.UK,
-                    deliveryClass = uk.gov.dluhc.printapi.database.entity.DeliveryClass.STANDARD,
-                    deliveryAddressType = DeliveryAddressType.ERO_COLLECTION,
-                    address = with(deliveryAddress) {
-                        buildAddress(
-                            street = street,
-                            postcode = postcode,
-                            property = property,
-                            locality = locality,
-                            town = town,
-                            area = area,
-                            uprn = uprn
-                        )
-                    },
-                )
-            }
-
-            // When
-            val actual = mapper.fromDeliveryDtoToDeliveryEntity(dto)
-
-            // Then
-            assertThat(actual).usingRecursiveComparison().ignoringFieldsMatchingRegexes(ID_FIELDS_REGEX).isEqualTo(expected)
-            verify(deliveryAddressTypeMapper).mapDtoToEntity(ERO_COLLECTION)
-        }
-    }
-
-    @Nested
-    inner class ToGenerateAnonymousElectorDocumentDto {
-
-        @Test
-        fun `should map to GenerateAnonymousElectorDocumentDto DTO given API Request`() {
-            // Given
-            val userId = aValidUserId()
-            val apiRequest = buildGenerateAnonymousElectorDocumentRequest(
-                sourceType = ANONYMOUS_MINUS_ELECTOR_MINUS_DOCUMENT,
-                certificateLanguage = uk.gov.dluhc.printapi.models.CertificateLanguage.EN,
-                supportingInformationFormat = AnonymousSupportingInformationFormatApi.LARGE_MINUS_PRINT,
-                delivery = buildApiCertificateDelivery(deliveryAddressType = ERO_MINUS_COLLECTION)
-            )
-
-            given(sourceTypeMapper.mapApiToDto(any())).willReturn(uk.gov.dluhc.printapi.dto.SourceType.ANONYMOUS_ELECTOR_DOCUMENT)
-            given(certificateLanguageMapper.mapApiToDto(any())).willReturn(uk.gov.dluhc.printapi.dto.CertificateLanguage.EN)
-            given(supportingInformationFormatMapper.mapApiToDto(any())).willReturn(AnonymousSupportingInformationFormatDto.LARGE_PRINT)
-            given(deliveryAddressTypeMapper.mapApiToDto(any())).willReturn(ERO_COLLECTION)
-
-            val expected = GenerateAnonymousElectorDocumentDto(
-                gssCode = apiRequest.gssCode,
-                sourceType = SourceType.ANONYMOUS_ELECTOR_DOCUMENT,
-                sourceReference = apiRequest.sourceReference,
-                applicationReference = apiRequest.applicationReference,
-                electoralRollNumber = apiRequest.electoralRollNumber,
-                photoLocation = apiRequest.photoLocation,
-                certificateLanguage = uk.gov.dluhc.printapi.dto.CertificateLanguage.EN,
-                supportingInformationFormat = AnonymousSupportingInformationFormatDto.LARGE_PRINT,
-                firstName = apiRequest.firstName,
-                middleNames = apiRequest.middleNames,
-                surname = apiRequest.surname,
-                email = apiRequest.email,
-                phoneNumber = apiRequest.phoneNumber,
-                registeredAddress = with(apiRequest.registeredAddress) {
-                    buildValidAddressDto(
-                        property = property,
-                        street = street,
-                        town = town,
-                        area = area,
-                        locality = locality,
-                        uprn = uprn,
-                        postcode = postcode
-                    )
-                },
-                userId = userId,
-                delivery = buildDtoCertificateDelivery(
-                    deliveryClass = DeliveryClass.STANDARD,
-                    deliveryAddressType = ERO_COLLECTION,
-                    addressee = apiRequest.delivery.addressee,
-                    addressFormat = AddressFormat.UK,
-                    deliveryAddress = with(apiRequest.delivery.deliveryAddress) {
+                    addressee = "Joe John Bloggs",
+                    registeredAddress = with(address!!) {
                         buildValidAddressDto(
-                            property = property,
-                            street = street,
-                            town = town,
-                            area = area,
-                            locality = locality,
-                            uprn = uprn,
-                            postcode = postcode
+                            street = street!!,
+                            postcode = postcode!!,
+                            property = property!!,
+                            locality = locality!!,
+                            town = town!!,
+                            area = area!!,
+                            uprn = uprn!!
                         )
-                    }
+                    },
+                    email = email,
+                    phoneNumber = phoneNumber
                 )
-            )
+            }
 
             // When
-            val actual = mapper.toGenerateAnonymousElectorDocumentDto(apiRequest, userId)
+            val actual = mapper.mapFromContactDetailsToElectorDto(aedContactDetailsEntity)
 
             // Then
-            assertThat(actual).isEqualTo(expected)
-            verify(sourceTypeMapper).mapApiToDto(ANONYMOUS_MINUS_ELECTOR_MINUS_DOCUMENT)
-            verify(certificateLanguageMapper).mapApiToDto(uk.gov.dluhc.printapi.models.CertificateLanguage.EN)
-            verify(supportingInformationFormatMapper).mapApiToDto(AnonymousSupportingInformationFormatApi.LARGE_MINUS_PRINT)
-            verify(deliveryAddressTypeMapper).mapApiToDto(ERO_MINUS_COLLECTION)
-            verifyNoMoreInteractions(sourceTypeMapper, certificateLanguageMapper, supportingInformationFormatMapper, deliveryAddressTypeMapper)
+            assertThat(actual).usingRecursiveComparison().isEqualTo(expected)
+        }
+
+        @Test
+        fun `should map AedContactDetails entity to registeredAddress of AnonymousElectorDto given only mandatory address fields present`() {
+            // Given
+            val aedContactDetailsEntity = buildAedContactDetails(
+                address = buildAddress(
+                    property = null, locality = null, town = null, area = null, uprn = null,
+                )
+            )
+            val expectedAddressDto = with(aedContactDetailsEntity.address!!) {
+                buildValidAddressDto(
+                    street = street!!,
+                    postcode = postcode!!,
+                    property = null, locality = null, town = null, area = null, uprn = null,
+                )
+            }
+
+            // When
+            val actual = mapper.mapFromContactDetailsToElectorDto(aedContactDetailsEntity)
+
+            // Then
+            assertThat(actual.registeredAddress).usingRecursiveComparison().isEqualTo(expectedAddressDto)
+        }
+
+        @ParameterizedTest
+        @NullAndEmptySource
+        @ValueSource(strings = ["  "])
+        fun `should map AedContactDetails entity names to Assignee for empty or null middleNames`(
+            middleNames: String?
+        ) {
+            // Given
+            val aedContactDetailsEntity =
+                buildAedContactDetails(firstName = "Joe", middleNames = middleNames, surname = "Bloggs")
+
+            // When
+            val actual = mapper.mapFromContactDetailsToElectorDto(aedContactDetailsEntity)
+
+            // Then
+            assertThat(actual.addressee).isEqualTo("Joe Bloggs")
         }
     }
 }
