+++ src/test/kotlin/uk/gov/dluhc/printapi/rest/aed/GetAnonymousElectorDocumentsByApplicationIdIntegrationTest.kt
@@ -8,7 +8,7 @@ import uk.gov.dluhc.printapi.database.entity.DeliveryAddressType.ERO_COLLECTION
 import uk.gov.dluhc.printapi.database.entity.DeliveryAddressType.REGISTERED
 import uk.gov.dluhc.printapi.database.entity.SourceType
 import uk.gov.dluhc.printapi.database.entity.SupportingInformationFormat
-import uk.gov.dluhc.printapi.models.AnonymousElectorDocumentSummariesResponse
+import uk.gov.dluhc.printapi.models.AnonymousElectorDocumentsResponse
 import uk.gov.dluhc.printapi.models.AnonymousSupportingInformationFormat.LARGE_MINUS_PRINT
 import uk.gov.dluhc.printapi.models.AnonymousSupportingInformationFormat.STANDARD
 import uk.gov.dluhc.printapi.models.DeliveryAddressType
@@ -21,13 +21,13 @@ import uk.gov.dluhc.printapi.testsupport.testdata.entity.buildDelivery
 import uk.gov.dluhc.printapi.testsupport.testdata.getRandomGssCode
 import uk.gov.dluhc.printapi.testsupport.testdata.getVCAnonymousAdminBearerToken
 import uk.gov.dluhc.printapi.testsupport.testdata.model.buildAnonymousElector
-import uk.gov.dluhc.printapi.testsupport.testdata.model.buildAnonymousElectorDocumentSummary
+import uk.gov.dluhc.printapi.testsupport.testdata.model.buildAnonymousElectorDocument
 import uk.gov.dluhc.printapi.testsupport.testdata.model.buildElectoralRegistrationOfficeResponse
 import uk.gov.dluhc.printapi.testsupport.testdata.model.buildLocalAuthorityResponse
 import uk.gov.dluhc.printapi.testsupport.testdata.model.buildValidAddress
 import java.time.ZoneOffset
 
-internal class GetAnonymousElectorDocumentSummaryByApplicationIdIntegrationTest : IntegrationTest() {
+internal class GetAnonymousElectorDocumentsByApplicationIdIntegrationTest : IntegrationTest() {
     companion object {
         private const val URI_TEMPLATE = "/eros/{ERO_ID}/anonymous-elector-documents?applicationId={APPLICATION_ID}"
         private const val APPLICATION_ID = "7762ccac7c056046b75d4bbc"
@@ -62,7 +62,7 @@ internal class GetAnonymousElectorDocumentSummaryByApplicationIdIntegrationTest 
     }
 
     @Test
-    fun `should return anonymous elector document summary`() {
+    fun `should return anonymous elector documents by application ID`() {
         // Given
         val eroResponse = buildElectoralRegistrationOfficeResponse(
             id = ERO_ID,
@@ -72,7 +72,9 @@ internal class GetAnonymousElectorDocumentSummaryByApplicationIdIntegrationTest 
         wireMockService.stubEroManagementGetEroByEroId(eroResponse, ERO_ID)
 
         val aedMatchingDocument1 = buildAnonymousElectorDocument(
-            gssCode = GSS_CODE, sourceReference = APPLICATION_ID, supportingInformationFormat = SupportingInformationFormat.LARGE_PRINT,
+            gssCode = GSS_CODE,
+            sourceReference = APPLICATION_ID,
+            supportingInformationFormat = SupportingInformationFormat.LARGE_PRINT,
             contactDetails = buildAedContactDetails(firstName = "John", middleNames = null, surname = "Jacob"),
             delivery = buildDelivery(deliveryAddressType = REGISTERED)
         )
@@ -82,7 +84,9 @@ internal class GetAnonymousElectorDocumentSummaryByApplicationIdIntegrationTest 
         Thread.sleep(1000)
 
         val aedMatchingDocument2 = buildAnonymousElectorDocument(
-            gssCode = GSS_CODE, sourceReference = APPLICATION_ID, supportingInformationFormat = SupportingInformationFormat.STANDARD,
+            gssCode = GSS_CODE,
+            sourceReference = APPLICATION_ID,
+            supportingInformationFormat = SupportingInformationFormat.STANDARD,
             contactDetails = buildAedContactDetails(firstName = "Mike", middleNames = "William Brown", surname = "Johnson"),
             delivery = buildDelivery(deliveryAddressType = ERO_COLLECTION)
         )
@@ -103,11 +107,14 @@ internal class GetAnonymousElectorDocumentSummaryByApplicationIdIntegrationTest 
         )
 
         val expectedFirstRecord = with(aedMatchingDocument2) {
-            buildAnonymousElectorDocumentSummary(
+            buildAnonymousElectorDocument(
                 certificateNumber = certificateNumber, electoralRollNumber = electoralRollNumber,
                 gssCode = gssCode, deliveryAddressType = DeliveryAddressType.ERO_MINUS_COLLECTION,
                 elector = with(contactDetails!!) {
                     buildAnonymousElector(
+                        firstName = firstName,
+                        middleNames = middleNames,
+                        surname = surname,
                         addressee = "Mike William Brown Johnson",
                         registeredAddress = with(address!!) {
                             buildValidAddress(
@@ -115,7 +122,9 @@ internal class GetAnonymousElectorDocumentSummaryByApplicationIdIntegrationTest 
                                 town = town, area = area, locality = locality,
                                 uprn = uprn, postcode = postcode!!
                             )
-                        }
+                        },
+                        email = email,
+                        phoneNumber = phoneNumber
                     )
                 },
                 photoLocation = photoLocationArn, issueDate = issueDate, userId = userId,
@@ -123,11 +132,14 @@ internal class GetAnonymousElectorDocumentSummaryByApplicationIdIntegrationTest 
             )
         }
         val expectedSecondRecord = with(aedMatchingDocument1) {
-            buildAnonymousElectorDocumentSummary(
+            buildAnonymousElectorDocument(
                 certificateNumber = certificateNumber, electoralRollNumber = electoralRollNumber,
                 gssCode = gssCode, deliveryAddressType = DeliveryAddressType.REGISTERED,
                 elector = with(contactDetails!!) {
                     buildAnonymousElector(
+                        firstName = firstName,
+                        middleNames = middleNames,
+                        surname = surname,
                         addressee = "John Jacob",
                         registeredAddress = with(address!!) {
                             buildValidAddress(
@@ -135,7 +147,9 @@ internal class GetAnonymousElectorDocumentSummaryByApplicationIdIntegrationTest 
                                 town = town, area = area, locality = locality,
                                 uprn = uprn, postcode = postcode!!
                             )
-                        }
+                        },
+                        email = email,
+                        phoneNumber = phoneNumber
                     )
                 },
                 photoLocation = photoLocationArn, issueDate = issueDate, userId = userId,
@@ -150,7 +164,7 @@ internal class GetAnonymousElectorDocumentSummaryByApplicationIdIntegrationTest 
             .contentType(APPLICATION_JSON)
             .exchange()
             .expectStatus().isOk
-            .returnResult(AnonymousElectorDocumentSummariesResponse::class.java)
+            .returnResult(AnonymousElectorDocumentsResponse::class.java)
 
         // Then
         val actual = response.responseBody.blockFirst()
