+++ src/test/kotlin/uk/gov/dluhc/printapi/service/aed/AnonymousElectorDocumentServiceTest.kt
@@ -20,13 +20,13 @@ import uk.gov.dluhc.printapi.database.entity.SourceType.ANONYMOUS_ELECTOR_DOCUME
 import uk.gov.dluhc.printapi.database.repository.AnonymousElectorDocumentRepository
 import uk.gov.dluhc.printapi.exception.GenerateAnonymousElectorDocumentValidationException
 import uk.gov.dluhc.printapi.mapper.aed.AnonymousElectorDocumentMapper
-import uk.gov.dluhc.printapi.mapper.aed.AnonymousElectorSummaryMapper
+import uk.gov.dluhc.printapi.mapper.aed.GenerateAnonymousElectorDocumentMapper
 import uk.gov.dluhc.printapi.service.EroService
 import uk.gov.dluhc.printapi.service.pdf.PdfFactory
 import uk.gov.dluhc.printapi.testsupport.testdata.aGssCode
 import uk.gov.dluhc.printapi.testsupport.testdata.aValidRandomEroId
 import uk.gov.dluhc.printapi.testsupport.testdata.aValidSourceReference
-import uk.gov.dluhc.printapi.testsupport.testdata.dto.aed.buildAnonymousElectorDocumentSummaryDto
+import uk.gov.dluhc.printapi.testsupport.testdata.dto.aed.buildAnonymousElectorDocumentDto
 import uk.gov.dluhc.printapi.testsupport.testdata.dto.aed.buildGenerateAnonymousElectorDocumentDto
 import uk.gov.dluhc.printapi.testsupport.testdata.entity.buildAnonymousElectorDocument
 import uk.gov.dluhc.printapi.testsupport.testdata.entity.buildDelivery
@@ -44,10 +44,10 @@ internal class AnonymousElectorDocumentServiceTest {
     private lateinit var anonymousElectorDocumentRepository: AnonymousElectorDocumentRepository
 
     @Mock
-    private lateinit var anonymousElectorDocumentMapper: AnonymousElectorDocumentMapper
+    private lateinit var generateAnonymousElectorDocumentMapper: GenerateAnonymousElectorDocumentMapper
 
     @Mock
-    private lateinit var anonymousElectorSummaryMapper: AnonymousElectorSummaryMapper
+    private lateinit var anonymousElectorDocumentMapper: AnonymousElectorDocumentMapper
 
     @Mock
     private lateinit var pdfTemplateDetailsFactory: AedPdfTemplateDetailsFactory
@@ -73,7 +73,7 @@ internal class AnonymousElectorDocumentServiceTest {
 
             given(eroService.isGssCodeValidForEro(any(), any())).willReturn(true)
             given(pdfTemplateDetailsFactory.getTemplateFilename(any())).willReturn(templateFilename)
-            given(anonymousElectorDocumentMapper.toAnonymousElectorDocument(any(), any())).willReturn(anonymousElectorDocument)
+            given(generateAnonymousElectorDocumentMapper.toAnonymousElectorDocument(any(), any())).willReturn(anonymousElectorDocument)
             given(pdfTemplateDetailsFactory.getTemplateDetails(any())).willReturn(templateDetails)
             given(pdfFactory.createPdfContents(any())).willReturn(contents)
 
@@ -83,7 +83,7 @@ internal class AnonymousElectorDocumentServiceTest {
             // Then
             verify(eroService).isGssCodeValidForEro(request.gssCode, eroId)
             verify(pdfTemplateDetailsFactory).getTemplateFilename(request.gssCode)
-            verify(anonymousElectorDocumentMapper).toAnonymousElectorDocument(request, templateFilename)
+            verify(generateAnonymousElectorDocumentMapper).toAnonymousElectorDocument(request, templateFilename)
             verify(pdfTemplateDetailsFactory).getTemplateDetails(anonymousElectorDocument)
             verify(pdfFactory).createPdfContents(templateDetails)
             verify(anonymousElectorDocumentRepository).save(anonymousElectorDocument)
@@ -108,7 +108,7 @@ internal class AnonymousElectorDocumentServiceTest {
             verify(eroService).isGssCodeValidForEro(request.gssCode, eroId)
             verifyNoInteractions(
                 pdfTemplateDetailsFactory,
-                anonymousElectorDocumentMapper,
+                generateAnonymousElectorDocumentMapper,
                 pdfTemplateDetailsFactory,
                 pdfFactory,
                 anonymousElectorDocumentRepository
@@ -133,7 +133,7 @@ internal class AnonymousElectorDocumentServiceTest {
             verify(eroService).isGssCodeValidForEro(request.gssCode, eroIdInRequest)
             verifyNoInteractions(
                 pdfTemplateDetailsFactory,
-                anonymousElectorDocumentMapper,
+                generateAnonymousElectorDocumentMapper,
                 pdfTemplateDetailsFactory,
                 pdfFactory,
                 anonymousElectorDocumentRepository
@@ -156,13 +156,13 @@ internal class AnonymousElectorDocumentServiceTest {
             given(anonymousElectorDocumentRepository.findByGssCodeInAndSourceTypeAndSourceReference(anyList(), any(), any())).willReturn(emptyList())
 
             // When
-            val actual = anonymousElectorDocumentService.getAnonymousElectorDocumentSummaries(eroId, applicationId)
+            val actual = anonymousElectorDocumentService.getAnonymousElectorDocuments(eroId, applicationId)
 
             // Then
             assertThat(actual).isNotNull.isEmpty()
             verify(eroService).lookupGssCodesForEro(eroId)
             verify(anonymousElectorDocumentRepository).findByGssCodeInAndSourceTypeAndSourceReference(gssCodes, ANONYMOUS_ELECTOR_DOCUMENT, applicationId)
-            verifyNoInteractions(anonymousElectorSummaryMapper)
+            verifyNoInteractions(anonymousElectorDocumentMapper)
             verifyNoMoreInteractions(eroService, anonymousElectorDocumentRepository)
         }
 
@@ -190,19 +190,19 @@ internal class AnonymousElectorDocumentServiceTest {
                     delivery = buildDelivery(deliveryAddressType = DeliveryAddressType.ERO_COLLECTION)
                 ).also { it.dateCreated = Instant.now().plusSeconds(2) }
 
-            val expectedDto1 = buildAnonymousElectorDocumentSummaryDto()
-            val expectedDto2 = buildAnonymousElectorDocumentSummaryDto()
-            val expectedDto3 = buildAnonymousElectorDocumentSummaryDto()
+            val expectedDto1 = buildAnonymousElectorDocumentDto()
+            val expectedDto2 = buildAnonymousElectorDocumentDto()
+            val expectedDto3 = buildAnonymousElectorDocumentDto()
 
             given(eroService.lookupGssCodesForEro(any())).willReturn(gssCodes)
             given(anonymousElectorDocumentRepository.findByGssCodeInAndSourceTypeAndSourceReference(anyList(), any(), any()))
                 .willReturn(listOf(firstAedEntity, secondAedEntity, aedEntityWithLatestDateCreated))
-            given(anonymousElectorSummaryMapper.mapToAnonymousElectorDocumentSummaryDto(firstAedEntity)).willReturn(expectedDto1)
-            given(anonymousElectorSummaryMapper.mapToAnonymousElectorDocumentSummaryDto(secondAedEntity)).willReturn(expectedDto2)
-            given(anonymousElectorSummaryMapper.mapToAnonymousElectorDocumentSummaryDto(aedEntityWithLatestDateCreated)).willReturn(expectedDto3)
+            given(anonymousElectorDocumentMapper.mapToAnonymousElectorDocumentDto(firstAedEntity)).willReturn(expectedDto1)
+            given(anonymousElectorDocumentMapper.mapToAnonymousElectorDocumentDto(secondAedEntity)).willReturn(expectedDto2)
+            given(anonymousElectorDocumentMapper.mapToAnonymousElectorDocumentDto(aedEntityWithLatestDateCreated)).willReturn(expectedDto3)
 
             // When
-            val actual = anonymousElectorDocumentService.getAnonymousElectorDocumentSummaries(eroId, applicationId)
+            val actual = anonymousElectorDocumentService.getAnonymousElectorDocuments(eroId, applicationId)
 
             // Then
             assertThat(actual).isNotNull.hasSize(3)
@@ -210,10 +210,10 @@ internal class AnonymousElectorDocumentServiceTest {
                 .isEqualTo(listOf(expectedDto3, expectedDto2, expectedDto1))
             verify(eroService).lookupGssCodesForEro(eroId)
             verify(anonymousElectorDocumentRepository).findByGssCodeInAndSourceTypeAndSourceReference(gssCodes, ANONYMOUS_ELECTOR_DOCUMENT, applicationId)
-            verify(anonymousElectorSummaryMapper).mapToAnonymousElectorDocumentSummaryDto(firstAedEntity)
-            verify(anonymousElectorSummaryMapper).mapToAnonymousElectorDocumentSummaryDto(secondAedEntity)
-            verify(anonymousElectorSummaryMapper).mapToAnonymousElectorDocumentSummaryDto(aedEntityWithLatestDateCreated)
-            verifyNoMoreInteractions(eroService, anonymousElectorDocumentRepository, anonymousElectorSummaryMapper)
+            verify(anonymousElectorDocumentMapper).mapToAnonymousElectorDocumentDto(firstAedEntity)
+            verify(anonymousElectorDocumentMapper).mapToAnonymousElectorDocumentDto(secondAedEntity)
+            verify(anonymousElectorDocumentMapper).mapToAnonymousElectorDocumentDto(aedEntityWithLatestDateCreated)
+            verifyNoMoreInteractions(eroService, anonymousElectorDocumentRepository, anonymousElectorDocumentMapper)
         }
     }
 }
