+++ src/test/kotlin/uk/gov/dluhc/printapi/testsupport/testdata/dto/aed/AnonymousElectorDocumentSummaryDtoBuilder.kt
@@ -3,8 +3,8 @@ package uk.gov.dluhc.printapi.testsupport.testdata.dto.aed
 import uk.gov.dluhc.printapi.dto.AddressDto
 import uk.gov.dluhc.printapi.dto.CertificateLanguage
 import uk.gov.dluhc.printapi.dto.DeliveryAddressType
+import uk.gov.dluhc.printapi.dto.aed.AnonymousElectorDocumentDto
 import uk.gov.dluhc.printapi.dto.aed.AnonymousElectorDocumentStatus
-import uk.gov.dluhc.printapi.dto.aed.AnonymousElectorDocumentSummaryDto
 import uk.gov.dluhc.printapi.dto.aed.AnonymousElectorDto
 import uk.gov.dluhc.printapi.dto.aed.AnonymousSupportingInformationFormat
 import uk.gov.dluhc.printapi.testsupport.testdata.aGssCode
@@ -12,14 +12,18 @@ import uk.gov.dluhc.printapi.testsupport.testdata.aValidApplicationReceivedDateT
 import uk.gov.dluhc.printapi.testsupport.testdata.aValidCertificateLanguageDto
 import uk.gov.dluhc.printapi.testsupport.testdata.aValidDeliveryName
 import uk.gov.dluhc.printapi.testsupport.testdata.aValidElectoralRollNumber
+import uk.gov.dluhc.printapi.testsupport.testdata.aValidEmailAddress
+import uk.gov.dluhc.printapi.testsupport.testdata.aValidFirstName
 import uk.gov.dluhc.printapi.testsupport.testdata.aValidIssueDate
+import uk.gov.dluhc.printapi.testsupport.testdata.aValidPhoneNumber
+import uk.gov.dluhc.printapi.testsupport.testdata.aValidSurname
 import uk.gov.dluhc.printapi.testsupport.testdata.aValidUserId
 import uk.gov.dluhc.printapi.testsupport.testdata.aValidVacNumber
 import uk.gov.dluhc.printapi.testsupport.testdata.zip.aPhotoArn
 import java.time.Instant
 import java.time.LocalDate
 
-fun buildAnonymousElectorDocumentSummaryDto(
+fun buildAnonymousElectorDocumentDto(
     certificateNumber: String = aValidVacNumber(),
     electoralRollNumber: String = aValidElectoralRollNumber(),
     gssCode: String = aGssCode(),
@@ -32,7 +36,7 @@ fun buildAnonymousElectorDocumentSummaryDto(
     issueDate: LocalDate = aValidIssueDate(),
     userId: String = aValidUserId(),
     requestDateTime: Instant = aValidApplicationReceivedDateTime()
-): AnonymousElectorDocumentSummaryDto = AnonymousElectorDocumentSummaryDto(
+): AnonymousElectorDocumentDto = AnonymousElectorDocumentDto(
     certificateNumber = certificateNumber,
     electoralRollNumber = electoralRollNumber,
     gssCode = gssCode,
@@ -48,9 +52,19 @@ fun buildAnonymousElectorDocumentSummaryDto(
 )
 
 fun buildAnonymousElectorDto(
+    firstName: String = aValidFirstName(),
+    middleNames: String? = null,
+    surname: String = aValidSurname(),
     addressee: String = aValidDeliveryName(),
-    registeredAddress: AddressDto = buildValidAddressDto()
+    registeredAddress: AddressDto = buildValidAddressDto(),
+    email: String? = aValidEmailAddress(),
+    phoneNumber: String? = aValidPhoneNumber()
 ): AnonymousElectorDto = AnonymousElectorDto(
+    firstName = firstName,
+    middleNames = middleNames,
+    surname = surname,
     addressee = addressee,
-    registeredAddress = registeredAddress
+    registeredAddress = registeredAddress,
+    email = email,
+    phoneNumber = phoneNumber
 )
